Started implementing a simple wizard on launch. It's still unusable
