+++ src/ChapterCreator.cpp
@@ -1,5 +1,5 @@
 #include "ChapterCreator.h"
-#include "Chapter.h"
+
 
 #include "MyApp.h"
