+++ src/MyApp.h
@@ -5,11 +5,13 @@
 
 #include <wx/wx.h>
 
-#include "Project.h"
+#include "ProjectManager.h"
+#include "ProjectWizard.h"
 
 class MyApp: public wxApp {
 private:
-    amdProjectManager* m_manager;
+    amdProjectManager* m_manager = nullptr;
+    amdProjectWizard* m_wizard = nullptr;
     wxLocale m_locale{};
 
 public:
@@ -17,6 +19,8 @@ public:
     virtual int OnExit();
 
     amdProjectManager* GetManager() { return m_manager; }
+    void OnWizardFinished(wxWizardEvent& event);
+    void OnWizardCanceled(wxWizardEvent& event);
 };
 
 wxDECLARE_APP(MyApp);
